+++ packages/streaming-image-volume-loader/src/BaseStreamingImageVolume.ts
@@ -177,6 +177,8 @@ export default class BaseStreamingImageVolume
     const eventDetail: Types.EventTypes.ImageVolumeModifiedEventDetail = {
       FrameOfReferenceUID,
       imageVolume: this,
+      numberOfFrames: numFrames,
+      framesProcessed: this.framesProcessed,
     };
 
     triggerEvent(eventTarget, Enums.Events.IMAGE_VOLUME_MODIFIED, eventDetail);
@@ -453,6 +455,7 @@ export default class BaseStreamingImageVolume
       },
       transferPixelData: true,
       transferSyntaxUID,
+      // The loader is used to load the image into the cache
       loader: imageLoader.loadImage,
       additionalDetails: {
         imageId,
@@ -663,7 +666,7 @@ export default class BaseStreamingImageVolume
     image,
     scalingParametersToUse: Types.ScalingParameters
   ) {
-    if (!image.preScale.enabled) {
+    if (!image.preScale?.enabled) {
       return image.getPixelData().slice(0);
     }