+++ packages/tools/src/synchronizers/synchronizers/createVOISynchronizer.ts
@@ -5,7 +5,7 @@ import Synchronizer from '../../store/SynchronizerManager/Synchronizer';
 
 type VOISynchronizerOptions = {
   syncInvertState: boolean;
-  syncColormap :boolean;
+  syncColormap: boolean;
 };
 
 /**
@@ -24,7 +24,10 @@ export default function createVOISynchronizer(
   options: VOISynchronizerOptions
 ): Synchronizer {
   //  = { syncInvertState: true } if options is not provided or undefined or {}
-  options = Object.assign({ syncInvertState: true, syncColormap:true }, options);
+  options = Object.assign(
+    { syncInvertState: true, syncColormap: true },
+    options
+  );
 
   const VOISynchronizer = createSynchronizer(
     synchronizerName,