+++ packages/tools/src/tools/TrackballRotateTool.ts
@@ -1,11 +1,16 @@
 import vtkMath from '@kitware/vtk.js/Common/Core/Math';
 import { Events } from '../enums';
 
-import { getEnabledElement } from '@cornerstonejs/core';
+import {
+  eventTarget,
+  getEnabledElement,
+  getEnabledElementByIds,
+} from '@cornerstonejs/core';
 import type { Types } from '@cornerstonejs/core';
 import { mat4, vec3 } from 'gl-matrix';
 import { EventTypes, PublicToolProps, ToolProps } from '../types';
 import { BaseTool } from './base';
+import { getToolGroup } from '../store/ToolGroupManager';
 
 /**
  * Tool that rotates the camera in the plane defined by the viewPlaneNormal and the viewUp.
@@ -15,6 +20,8 @@ class TrackballRotateTool extends BaseTool {
   touchDragCallback: (evt: EventTypes.InteractionEventType) => void;
   mouseDragCallback: (evt: EventTypes.InteractionEventType) => void;
   cleanUp: () => void;
+  _resizeObservers = new Map();
+  _viewportAddedListener: (evt: any) => void;
 
   constructor(
     toolProps: PublicToolProps = {},
@@ -56,6 +63,77 @@ class TrackballRotateTool extends BaseTool {
 
     document.addEventListener('mouseup', this.cleanUp, { once: true });
     return true;
+  };
+
+  _getViewportsInfo = () => {
+    const viewports = getToolGroup(this.toolGroupId).viewportsInfo;
+
+    return viewports;
+  };
+
+  onSetToolActive = () => {
+    const subscribeToElementResize = () => {
+      const viewportsInfo = this._getViewportsInfo();
+      viewportsInfo.forEach(({ viewportId, renderingEngineId }) => {
+        if (!this._resizeObservers.has(viewportId)) {
+          const { viewport } = getEnabledElementByIds(
+            viewportId,
+            renderingEngineId
+          ) || { viewport: null };
+
+          if (!viewport) {
+            return;
+          }
+
+          const { element } = viewport;
+
+          const resizeObserver = new ResizeObserver(() => {
+            const element = getEnabledElementByIds(
+              viewportId,
+              renderingEngineId
+            );
+            if (!element) {
+              return;
+            }
+            const { viewport } = element;
+            viewport.resetCamera();
+            viewport.render();
+          });
+
+          resizeObserver.observe(element);
+          this._resizeObservers.set(viewportId, resizeObserver);
+        }
+      });
+    };
+
+    subscribeToElementResize();
+
+    this._viewportAddedListener = (evt) => {
+      if (evt.detail.toolGroupId === this.toolGroupId) {
+        subscribeToElementResize();
+      }
+    };
+
+    eventTarget.addEventListener(
+      Events.TOOLGROUP_VIEWPORT_ADDED,
+      this._viewportAddedListener
+    );
+  };
+
+  onSetToolDisabled = () => {
+    // Disconnect all resize observers
+    this._resizeObservers.forEach((resizeObserver, viewportId) => {
+      resizeObserver.disconnect();
+      this._resizeObservers.delete(viewportId);
+    });
+
+    if (this._viewportAddedListener) {
+      eventTarget.removeEventListener(
+        Events.TOOLGROUP_VIEWPORT_ADDED,
+        this._viewportAddedListener
+      );
+      this._viewportAddedListener = null; // Clear the reference to the listener
+    }
   };
 
   rotateCamera = (viewport, centerWorld, axis, angle) => {